Move none_like out of util to fix galaxy.util dependencies.
