+++ lib/galaxy/tools/evaluation.py
@@ -8,6 +8,7 @@ from six import string_types
 
 from galaxy import model
 from galaxy.jobs.datasets import dataset_path_rewrites
+from galaxy.model.none_like import NoneDataset
 from galaxy.tools import global_tool_errors
 from galaxy.tools.parameters import (
     visit_input_values,
@@ -38,7 +39,6 @@ from galaxy.util import (
     unicodify,
 )
 from galaxy.util.bunch import Bunch
-from galaxy.util.none_like import NoneDataset
 from galaxy.util.object_wrapper import wrap_with_safe_string
 from galaxy.util.template import fill_template
 from galaxy.work.context import WorkRequestContext
